refactor bundle versions and fraud proof versions to specific modules
