+++ crates/pallet-domains/src/benchmarking.rs
@@ -34,14 +34,15 @@ use sp_consensus_slots::Slot;
 use sp_core::H256;
 use sp_core::crypto::{Ss58Codec, UncheckedFrom};
 use sp_core::sr25519::vrf::{VrfPreOutput, VrfProof, VrfSignature};
+use sp_domains::bundle::bundle_v1::BundleV1;
+use sp_domains::bundle::{BundleHeader, SealedBundleHeader, dummy_opaque_bundle};
 use sp_domains::merkle_tree::MerkleTree;
 use sp_domains::{
-    BundleHeader, BundleV1, DomainId, EMPTY_EXTRINSIC_ROOT, ExecutionReceipt, OperatorAllowList,
-    OperatorId, OperatorPublicKey, OperatorRewardSource, OperatorSignature,
-    PermissionedActionAllowedBy, ProofOfElection, RuntimeType, SealedBundleHeader,
-    SealedSingletonReceipt, SingletonReceipt, dummy_opaque_bundle,
+    DomainId, EMPTY_EXTRINSIC_ROOT, ExecutionReceipt, OperatorAllowList, OperatorId,
+    OperatorPublicKey, OperatorRewardSource, OperatorSignature, PermissionedActionAllowedBy,
+    ProofOfElection, RuntimeType, SealedSingletonReceipt, SingletonReceipt,
 };
-use sp_domains_fraud_proof::fraud_proof_v1::FraudProofV1;
+use sp_domains_fraud_proof::fraud_proof::fraud_proof_v1::FraudProofV1;
 use sp_runtime::traits::{CheckedAdd, One, Zero};
 use sp_std::collections::btree_set::BTreeSet;
 use subspace_core_primitives::Randomness;
@@ -967,7 +968,7 @@ mod benchmarks {
             201, 155, 176, 188, 254, 114, 173, 96, 134,
         ]);
 
-        let opaque_bundle = VersionedOpaqueBundle::V1(BundleV1 {
+        let opaque_bundle = OpaqueBundle::V1(BundleV1 {
             sealed_header: SealedBundleHeader::new(header, signature),
             extrinsics: Vec::new(),
         });
