+++ crates/sp-domains-fraud-proof/src/fraud_proof/fraud_proof_v0.rs
@@ -11,7 +11,8 @@ use alloc::vec::Vec;
 use core::fmt;
 use parity_scale_codec::{Decode, Encode};
 use scale_info::TypeInfo;
-use sp_domains::{DomainId, HeaderHashFor, HeaderHashingFor, InvalidBundleType};
+use sp_domains::bundle::InvalidBundleType;
+use sp_domains::{DomainId, HeaderHashFor, HeaderHashingFor};
 use sp_runtime::traits::{Hash as HashT, Header as HeaderT};
 use sp_subspace_mmr::ConsensusChainMmrLeafProof;
 use sp_trie::StorageProof;
