+++ crates/sp-domains-fraud-proof/src/lib.rs
@@ -9,8 +9,6 @@
 #[cfg(feature = "std")]
 pub mod execution_prover;
 pub mod fraud_proof;
-pub mod fraud_proof_v0;
-pub mod fraud_proof_v1;
 #[cfg(feature = "std")]
 mod host_functions;
 mod runtime_interface;
@@ -23,12 +21,12 @@ pub mod weights;
 #[cfg(not(feature = "std"))]
 extern crate alloc;
 
-use crate::fraud_proof_v0::FraudProofV0;
-use crate::fraud_proof_v1::FraudProofV1;
 use crate::storage_proof::FraudProofStorageKeyRequest;
 #[cfg(not(feature = "std"))]
 use alloc::vec::Vec;
 use domain_runtime_primitives::EthereumAccountId;
+use fraud_proof::fraud_proof_v0::FraudProofV0;
+use fraud_proof::fraud_proof_v1::FraudProofV1;
 #[cfg(feature = "std")]
 pub use host_functions::{
     FraudProofExtension, FraudProofHostFunctions, FraudProofHostFunctionsImpl,
