+++ crates/sp-domains-fraud-proof/src/verification.rs
@@ -1,13 +1,13 @@
 #[cfg(not(feature = "std"))]
 extern crate alloc;
 
-use crate::fraud_proof::{
-    InvalidExtrinsicsRootProof, InvalidStateTransitionProof, MmrRootProof, VerificationError,
-};
-use crate::fraud_proof_v0::{InvalidBundlesV0Proof, ValidBundleV0Proof};
-use crate::fraud_proof_v1::{
+use crate::fraud_proof::fraud_proof_v0::{InvalidBundlesV0Proof, ValidBundleV0Proof};
+use crate::fraud_proof::fraud_proof_v1::{
     InvalidBundlesProof, InvalidBundlesProofData, InvalidVersionedBundlesProof, ValidBundleProof,
     ValidVersionedBundleProof,
+};
+use crate::fraud_proof::{
+    InvalidExtrinsicsRootProof, InvalidStateTransitionProof, MmrRootProof, VerificationError,
 };
 use crate::storage_proof::{self, *};
 use crate::{
@@ -21,13 +21,13 @@ use hash_db::Hasher;
 use parity_scale_codec::{Decode, Encode};
 use sp_core::H256;
 use sp_core::storage::StorageKey;
+use sp_domains::bundle::{BundleValidity, InboxedBundle, InvalidBundleType};
 use sp_domains::extrinsics::deduplicate_and_shuffle_extrinsics;
 use sp_domains::proof_provider_and_verifier::StorageProofVerifier;
 use sp_domains::valued_trie::valued_ordered_trie_root;
 use sp_domains::{
-    BlockFees, BundleValidity, DomainId, ExecutionReceipt, ExtrinsicDigest, HeaderHashFor,
-    HeaderHashingFor, HeaderNumberFor, INITIAL_DOMAIN_TX_RANGE, InboxedBundle, InvalidBundleType,
-    RuntimeId, Transfers,
+    BlockFees, DomainId, ExecutionReceipt, ExtrinsicDigest, HeaderHashFor, HeaderHashingFor,
+    HeaderNumberFor, INITIAL_DOMAIN_TX_RANGE, RuntimeId, Transfers,
 };
 use sp_runtime::generic::Digest;
 use sp_runtime::traits::{
