+++ crates/subspace-fake-runtime-api/src/lib.rs
@@ -8,12 +8,13 @@ use frame_support::weights::Weight;
 use sp_consensus_subspace::{ChainConstants, PotParameters, SignedVote, SolutionRanges};
 use sp_core::crypto::KeyTypeId;
 use sp_core::{H256, OpaqueMetadata};
+use sp_domains::bundle::BundleVersion;
 use sp_domains::bundle_producer_election::BundleProducerElectionParams;
 use sp_domains::{
-    BundleVersion, DomainAllowlistUpdates, DomainId, DomainInstanceData, ExecutionReceiptFor,
-    OperatorId, OperatorPublicKey, PermissionedActionAllowedBy,
+    DomainAllowlistUpdates, DomainId, DomainInstanceData, ExecutionReceiptFor, OperatorId,
+    OperatorPublicKey, PermissionedActionAllowedBy,
 };
-use sp_domains_fraud_proof::fraud_proof_v1::FraudProofV1;
+use sp_domains_fraud_proof::fraud_proof::fraud_proof_v1::FraudProofV1;
 use sp_domains_fraud_proof::storage_proof::FraudProofStorageKeyRequest;
 use sp_messenger::messages::{
     BlockMessagesQuery, BlockMessagesWithStorageKey, ChainId, ChannelId, ChannelStateWithNonce,
@@ -169,7 +170,7 @@ sp_api::impl_runtime_apis! {
 
     impl sp_domains::DomainsApi<Block, DomainHeader> for Runtime {
         fn submit_bundle_unsigned(
-            _opaque_bundle: sp_domains::VersionedOpaqueBundle<NumberFor<Block>, BlockHashFor<Block>, DomainHeader, Balance>,
+            _opaque_bundle: sp_domains::bundle::OpaqueBundle<NumberFor<Block>, BlockHashFor<Block>, DomainHeader, Balance>,
         ) {
             unreachable!()
         }
@@ -183,7 +184,7 @@ sp_api::impl_runtime_apis! {
         fn extract_successful_bundles(
             _domain_id: DomainId,
             _extrinsics: Vec<ExtrinsicFor<Block>>,
-        ) -> sp_domains::VersionedOpaqueBundles<Block, DomainHeader, Balance> {
+        ) -> sp_domains::bundle::VersionedOpaqueBundles<Block, DomainHeader, Balance> {
             unreachable!()
         }
