+++ crates/subspace-runtime/src/lib.rs
@@ -58,13 +58,14 @@ use sp_consensus_slots::{Slot, SlotDuration};
 use sp_consensus_subspace::{ChainConstants, PotParameters, SignedVote, SolutionRanges, Vote};
 use sp_core::crypto::KeyTypeId;
 use sp_core::{ConstBool, H256, OpaqueMetadata};
+use sp_domains::bundle::BundleVersion;
 use sp_domains::bundle_producer_election::BundleProducerElectionParams;
 use sp_domains::{
-    BundleVersion, ChannelId, DOMAIN_STORAGE_FEE_MULTIPLIER, DomainAllowlistUpdates, DomainId,
-    DomainInstanceData, ExecutionReceiptFor, INITIAL_DOMAIN_TX_RANGE, OperatorId,
-    OperatorPublicKey, PermissionedActionAllowedBy,
+    ChannelId, DOMAIN_STORAGE_FEE_MULTIPLIER, DomainAllowlistUpdates, DomainId, DomainInstanceData,
+    ExecutionReceiptFor, INITIAL_DOMAIN_TX_RANGE, OperatorId, OperatorPublicKey,
+    PermissionedActionAllowedBy,
 };
-use sp_domains_fraud_proof::fraud_proof_v1::FraudProofV1;
+use sp_domains_fraud_proof::fraud_proof::fraud_proof_v1::FraudProofV1;
 use sp_domains_fraud_proof::storage_proof::{
     FraudProofStorageKeyProvider, FraudProofStorageKeyRequest,
 };
@@ -1473,7 +1474,7 @@ impl_runtime_apis! {
 
     impl sp_domains::DomainsApi<Block, DomainHeader> for Runtime {
         fn submit_bundle_unsigned(
-            opaque_bundle: sp_domains::VersionedOpaqueBundle<NumberFor<Block>, BlockHashFor<Block>, DomainHeader, Balance>,
+            opaque_bundle: sp_domains::bundle::OpaqueBundle<NumberFor<Block>, BlockHashFor<Block>, DomainHeader, Balance>,
         ) {
             Domains::submit_bundle_unsigned(opaque_bundle)
         }
@@ -1487,7 +1488,7 @@ impl_runtime_apis! {
         fn extract_successful_bundles(
             domain_id: DomainId,
             extrinsics: Vec<ExtrinsicFor<Block>>,
-        ) -> sp_domains::VersionedOpaqueBundles<Block, DomainHeader, Balance> {
+        ) -> sp_domains::bundle::VersionedOpaqueBundles<Block, DomainHeader, Balance> {
             crate::domains::extract_successful_bundles(domain_id, extrinsics)
         }
