+++ domains/client/domain-operator/src/aux_schema.rs
@@ -4,7 +4,7 @@ use crate::ExecutionReceiptFor;
 use parity_scale_codec::{Decode, Encode};
 use sc_client_api::backend::AuxStore;
 use sp_blockchain::{Error as ClientError, HeaderBackend, Result as ClientResult};
-use sp_domains::InvalidBundleType;
+use sp_domains::bundle::InvalidBundleType;
 use sp_runtime::Saturating;
 use sp_runtime::traits::{
     Block as BlockT, CheckedMul, CheckedSub, NumberFor, One, SaturatedConversion, Zero,
