+++ domains/client/domain-operator/src/lib.rs
@@ -101,7 +101,8 @@ use sp_blockchain::HeaderBackend;
 use sp_consensus::SyncOracle;
 use sp_consensus_slots::Slot;
 use sp_domain_digests::AsPredigest;
-use sp_domains::{BundleV1, DomainId, ExecutionReceiptFor as ExecutionReceipt, OperatorId};
+use sp_domains::bundle::bundle_v1::BundleV1;
+use sp_domains::{DomainId, ExecutionReceiptFor as ExecutionReceipt, OperatorId};
 use sp_keystore::KeystorePtr;
 use sp_runtime::DigestItem;
 use sp_runtime::traits::{Block as BlockT, Header as HeaderT, NumberFor};
