+++ domains/client/domain-operator/src/tests.rs
@@ -39,18 +39,21 @@ use sp_core::storage::StateVersion;
 use sp_core::traits::{FetchRuntimeCode, SpawnEssentialNamed};
 use sp_core::{H160, H256, Pair, U256};
 use sp_domain_digests::AsPredigest;
+use sp_domains::bundle::{Bundle, BundleValidity, InboxedBundle, InvalidBundleType};
 use sp_domains::core_api::DomainCoreApi;
 use sp_domains::merkle_tree::MerkleTree;
 use sp_domains::{
-    BlockFees, BundleValidity, ChainId, ChannelId, DomainsApi, HeaderHashingFor, InboxedBundle,
-    InvalidBundleType, OperatorPublicKey, PermissionedActionAllowedBy, Transfers, VersionedBundle,
+    BlockFees, ChainId, ChannelId, DomainsApi, HeaderHashingFor, OperatorPublicKey,
+    PermissionedActionAllowedBy, Transfers,
 };
 use sp_domains_fraud_proof::InvalidTransactionCode;
+use sp_domains_fraud_proof::fraud_proof::fraud_proof_v1::{
+    FraudProofVariantV1, InvalidBundlesProofData,
+};
 use sp_domains_fraud_proof::fraud_proof::{
     ApplyExtrinsicMismatch, ExecutionPhase, FinalizeBlockMismatch, InvalidBlockFeesProof,
     InvalidDomainBlockHashProof, InvalidExtrinsicsRootProof, InvalidTransfersProof,
 };
-use sp_domains_fraud_proof::fraud_proof_v1::{FraudProofVariantV1, InvalidBundlesProofData};
 use sp_messenger::MessengerApi;
 use sp_messenger::messages::{CrossDomainMessage, Proof};
 use sp_mmr_primitives::{EncodableOpaqueLeaf, LeafProof as MmrProof};
@@ -1892,7 +1895,7 @@ async fn collected_receipts_should_be_on_the_same_branch_with_current_best_block
     let consensus_block_info =
         |best_header: Header| -> (u32, Hash) { (*best_header.number(), best_header.hash()) };
     let receipts_consensus_info =
-        |bundle: VersionedBundle<OpaqueExtrinsic, u32, sp_core::H256, Header, Balance>| {
+        |bundle: Bundle<OpaqueExtrinsic, u32, sp_core::H256, Header, Balance>| {
             (
                 bundle.receipt().consensus_block_number,
                 bundle.receipt().consensus_block_hash,
