+++ test/subspace-test-runtime/src/lib.rs
@@ -70,13 +70,14 @@ use sp_consensus_slots::{Slot, SlotDuration};
 use sp_consensus_subspace::{ChainConstants, PotParameters, SignedVote, SolutionRanges, Vote};
 use sp_core::crypto::KeyTypeId;
 use sp_core::{H256, OpaqueMetadata};
+use sp_domains::bundle::{BundleVersion, OpaqueBundle, VersionedOpaqueBundles};
 use sp_domains::bundle_producer_election::BundleProducerElectionParams;
 use sp_domains::{
-    BundleVersion, DOMAIN_STORAGE_FEE_MULTIPLIER, DomainAllowlistUpdates, DomainId,
-    DomainInstanceData, ExecutionReceiptFor, INITIAL_DOMAIN_TX_RANGE, OperatorId,
-    OperatorPublicKey, PermissionedActionAllowedBy, VersionedOpaqueBundle, VersionedOpaqueBundles,
+    DOMAIN_STORAGE_FEE_MULTIPLIER, DomainAllowlistUpdates, DomainId, DomainInstanceData,
+    ExecutionReceiptFor, INITIAL_DOMAIN_TX_RANGE, OperatorId, OperatorPublicKey,
+    PermissionedActionAllowedBy,
 };
-use sp_domains_fraud_proof::fraud_proof_v1::FraudProofV1;
+use sp_domains_fraud_proof::fraud_proof::fraud_proof_v1::FraudProofV1;
 use sp_domains_fraud_proof::storage_proof::{
     FraudProofStorageKeyProvider, FraudProofStorageKeyRequest,
 };
@@ -1513,7 +1514,7 @@ impl_runtime_apis! {
 
     impl sp_domains::DomainsApi<Block, DomainHeader> for Runtime {
         fn submit_bundle_unsigned(
-            opaque_bundle: VersionedOpaqueBundle<NumberFor<Block>, BlockHashFor<Block>, DomainHeader, Balance>,
+            opaque_bundle: OpaqueBundle<NumberFor<Block>, BlockHashFor<Block>, DomainHeader, Balance>,
         ) {
             Domains::submit_bundle_unsigned(opaque_bundle)
         }
